// src/extension_core/wcp/generated/wcp_types.ts
export type ItemPath = string;

export interface MarkerInfo {
  time: number;
  name?: string;
  move_focus: boolean;
}

export interface ItemInfo {
  name: string;
  type: string;
  id: DisplayedItemRef;
}

export interface WcpGreeting {
  type: "greeting";
  version: string;
  commands: Array<string>;
}

export interface WcpError {
  type: "error";
  message?: string;
}

export interface GetItemListCommand {
  type: "command";
  command: "get_item_list";
}

export interface GetItemInfoParams {
  ids: Array<DisplayedItemRef>;
}

export interface GetItemInfoCommand extends GetItemInfoParams {
  type: "command";
  command: "get_item_info";
}

export interface SetItemColorParams {
  id: DisplayedItemRef;
  color: string;
}

export interface SetItemColorCommand extends SetItemColorParams {
  type: "command";
  command: "set_item_color";
}

export interface AddItemsParams {
  items: Array<ItemPath>;
  recursive: boolean;
}

export interface AddItemsCommand extends AddItemsParams {
  type: "command";
  command: "add_items";
}

export interface RemoveItemsParams {
  ids: Array<DisplayedItemRef>;
}

export interface RemoveItemsCommand extends RemoveItemsParams {
  type: "command";
  command: "remove_items";
}

export interface FocusItemParams {
  id: DisplayedItemRef;
}

export interface FocusItemCommand extends FocusItemParams {
  type: "command";
  command: "focus_item";
}

export interface AddMarkersParams {
  items: Array<MarkerInfo>;
}

export interface AddMarkersCommand extends AddMarkersParams {
  type: "command";
  command: "add_markers";
}

export interface SetViewportToParams {
  timestamp: number;
}

export interface SetViewportToCommand extends SetViewportToParams {
  type: "command";
  command: "set_viewport_to";
}

export interface SetViewportRangeParams {
  start: number;
  end: number;
}

export interface SetViewportRangeCommand extends SetViewportRangeParams {
  type: "command";
  command: "set_viewport_range";
}

export interface ZoomToFitCommand {
  type: "command";
  command: "zoom_to_fit";
}

export interface SetCursorParams {
  timestamp: number;
}

export interface SetCursorCommand extends SetCursorParams {
  type: "command";
  command: "set_cursor";
}

export interface LoadParams {
  source: string;
}

export interface LoadCommand extends LoadParams {
  type: "command";
  command: "load";
}

export interface ReloadCommand {
  type: "command";
  command: "reload";
}

export interface ClearCommand {
  type: "command";
  command: "clear";
}

export interface ShutdownCommand {
  type: "command";
  command: "shutdown";
}

export interface GetItemListResponse {
  type: "response";
  command: "get_item_list";
  ids: Array<DisplayedItemRef>;
}

export interface GetItemInfoResponse {
  type: "response";
  command: "get_item_info";
  results: Array<ItemInfo>;
}

export interface AddItemsResponse {
  type: "response";
  command: "add_items";
  ids: Array<DisplayedItemRef>;
}

export interface AddMarkersResponse {
  type: "response";
  command: "add_markers";
  ids: Array<DisplayedItemRef>;
}

export interface AckResponse {
  type: "response";
  command: "ack";
}

export interface WaveformsLoadedEvent {
  type: "event";
  event: "waveforms_loaded";
  source: string;
}

export interface CursorSetEvent {
  type: "event";
  event: "cursor_set";
  timestamp: number;
}

export interface GotoDeclarationEvent {
  type: "event";
  event: "goto_declaration";
  variable: ItemPath;
}

export interface AddDriversEvent {
  type: "event";
  event: "add_drivers";
  variable: ItemPath;
}

export interface AddLoadsEvent {
  type: "event";
  event: "add_loads";
  variable: ItemPath;
}

export type WcpCommand =
  | GetItemListCommand
  | GetItemInfoCommand
  | SetItemColorCommand
  | AddItemsCommand
  | RemoveItemsCommand
  | FocusItemCommand
  | AddMarkersCommand
  | SetViewportToCommand
  | SetViewportRangeCommand
  | ZoomToFitCommand
  | SetCursorCommand
  | LoadCommand
  | ReloadCommand
  | ClearCommand
  | ShutdownCommand;

export type WcpResponse =
  | GetItemListResponse
  | GetItemInfoResponse
  | AddItemsResponse
  | AddMarkersResponse
  | AckResponse;

export type WcpEvent =
  | WaveformsLoadedEvent
  | CursorSetEvent
  | GotoDeclarationEvent
  | AddDriversEvent
  | AddLoadsEvent;
// ...
export type WcpMessage = WcpGreeting | WcpCommand | WcpResponse | WcpEvent | WcpError;